+++ praw/models/reddit/live.py
@@ -431,3 +431,9 @@ class LiveUpdateContribution(object):
         url = API_PATH['live_remove_update'].format(id=self.update.thread.id)
         data = {'id': self.update.fullname}
         self.update.thread._reddit.post(url, data=data)
+
+    def strike(self):
+        """Strike a content of a live update."""
+        url = API_PATH['live_strike'].format(id=self.update.thread.id)
+        data = {'id': self.update.fullname}
+        self.update.thread._reddit.post(url, data=data)