+++ CHANGES.rst
@@ -13,6 +13,7 @@ Unreleased
 * :meth:`.upload_mobile_icon` to upload subreddit mobile icon.
 * :meth:`.delete_mobile_header` to remove subreddit mobile header.
 * :meth:`.delete_mobile_icon` to remove subreddit mobile icon.
+* :meth:`.LiveUpdateContribution.strike` to strike a content of a live thread.
 
 **Fixed**